Turning tracks.json into album-level metadata.
Previously, tracks.json was a list of track metadata, nothing more.

This had the unintended consequence of encoding the image for each track in the album multiple times.

Now, we take the first image we find, and bring that up to the album level.

While we're at it, we collect the rest of the album-level metadata here as well, so that tracks.json now looks like

```
{ album, artist, image, tracks }
```

Additionally, the programmatic usage now calls back on not only the ipfs hash, but also the `tracks` data we created.

Turning tracks.json into album-level metadata.

refactor