Single script for adding music folders
Previously, we were adding directories of music -- and generated, associated metadata -- to IPFS. Then, separately, we were downloading the metadata file from IPFS, using it to generate *more* (album-level) metadata, then adding the metadata (and IPFS hash) to a sqlite db.

Since the prior commit, we generate album-level metadata in the file we upload to IPFS, and call back on this metadata.

In this commit, we introduce a single script that will add music files to IPFS, generate and add associated metadata, and directly add the generated metadata to a SQLite db, without fetching anything from IPFS between these steps.

Removed library.db

..